+++ src/tests/test_all.py
@@ -12,35 +12,13 @@ import os
 import shutil
 import sys
 from tabfix import main, cmd_walker
-#from unittest.suite import TestSuite
 from tabfix.main import read_text_lines, DELIM_CR, DELIM_CRLF, DELIM_LF
-
-IS_PY3 = (sys.version_info[0] >= 3) 
+import subprocess
+import StringIO
+#IS_PY3 = (sys.version_info[0] >= 3) 
 
 USE_FIXED_FOLDER = False
 
-#class TextFileReader(object):
-#    def __init__(self, fname):
-#        if sys.version_info[0] >= 3:
-#            # Python 3: read binary, because we don't know the encoding.
-#            # We still want line-breaks, so U is added.
-#            # NOTE: '\n' and '\r\n' are recognized, but '\r' NOT
-#            mode = "Urb"
-#            self.f = open(fname, "Urb")
-#        else:
-#            # Python 2: we get 
-#            self.f = open(fname, "Urt")
-#    def read_text_as_binary(self, fname):
-#        """"""
-#        is_py3 = (sys.version_info[0] >= 3) 
-#        if is_py3:
-#            # Python 3: read binary, because we don't know the encoding.
-#            # We still want line-breaks, so U is added.
-#            # NOTE: '\n' and '\r\n' are recognized, but '\r' NOT
-#            mode = "Urb"
-#        else:
-#            # Python 2: we get 
-#            mode = "Urt"
 
 class TestBasic(unittest.TestCase):
     """Basic tests.
@@ -73,6 +51,40 @@ class TestBasic(unittest.TestCase):
     def tearDown(self):
         os.chdir(self.prev_cwd)
         shutil.rmtree(self.temp_path)
+        
+    def test_read_text_lines(self):
+        
+        stats = { DELIM_CR: 0, DELIM_LF: 0, DELIM_CRLF: 0 }
+        res = read_text_lines("test_cr.txt", stats)
+        res = list(res)
+        self.assertEqual(len(res), 7)
+        self.assertEqual(type(res[0]), type(b""))
+        self.assertTrue(res[0].endswith(DELIM_CR))
+        self.assertEqual(stats, { DELIM_CR: 6, DELIM_LF: 0, DELIM_CRLF: 0 })
+        
+        stats = { DELIM_CR: 0, DELIM_LF: 0, DELIM_CRLF: 0 }
+        res = read_text_lines("test_crlf.txt", stats)
+        res = list(res)
+        self.assertEqual(len(res), 7)
+        self.assertEqual(type(res[0]), type(b""))
+        self.assertTrue(res[0].endswith(DELIM_CRLF))
+        self.assertEqual(stats, { DELIM_CR: 0, DELIM_LF: 0, DELIM_CRLF: 6 })
+        
+        stats = { DELIM_CR: 0, DELIM_LF: 0, DELIM_CRLF: 0 }
+        res = read_text_lines("test_lf.txt", stats)
+        res = list(res)
+        self.assertEqual(len(res), 7)
+        self.assertEqual(type(res[0]), type(b""))
+        self.assertTrue(res[0].endswith(DELIM_LF))
+        self.assertEqual(stats, { DELIM_CR: 0, DELIM_LF: 6, DELIM_CRLF: 0 })
+        
+        stats = { DELIM_CR: 0, DELIM_LF: 0, DELIM_CRLF: 0 }
+        res = read_text_lines("test_mixed.txt", stats)
+        res = list(res)
+        self.assertEqual(len(res), 45)
+        self.assertEqual(type(res[0]), type(b""))
+        self.assertTrue(res[0].endswith(DELIM_CRLF))
+        self.assertEqual(stats, { DELIM_CR: 0, DELIM_LF: 0, DELIM_CRLF: 44 })
         
     def test_spacify_txt_flat(self):
         
@@ -178,47 +190,63 @@ class TestBasic(unittest.TestCase):
         self.assertEqual(data.get("files_processed"), 16)
 
 
-    def test_read_text_lines(self):
-        
-        stats = { DELIM_CR: 0, DELIM_LF: 0, DELIM_CRLF: 0 }
-        res = read_text_lines("test_cr.txt", stats)
-        res = list(res)
-        self.assertEqual(len(res), 7)
-        self.assertEqual(type(res[0]), type(b""))
-        self.assertTrue(res[0].endswith(DELIM_CR))
-        self.assertEqual(stats, { DELIM_CR: 6, DELIM_LF: 0, DELIM_CRLF: 0 })
-        
-        stats = { DELIM_CR: 0, DELIM_LF: 0, DELIM_CRLF: 0 }
-        res = read_text_lines("test_crlf.txt", stats)
-        res = list(res)
-        self.assertEqual(len(res), 7)
-        self.assertEqual(type(res[0]), type(b""))
-        self.assertTrue(res[0].endswith(DELIM_CRLF))
-        self.assertEqual(stats, { DELIM_CR: 0, DELIM_LF: 0, DELIM_CRLF: 6 })
-        
-        stats = { DELIM_CR: 0, DELIM_LF: 0, DELIM_CRLF: 0 }
-        res = read_text_lines("test_lf.txt", stats)
-        res = list(res)
-        self.assertEqual(len(res), 7)
-        self.assertEqual(type(res[0]), type(b""))
-        self.assertTrue(res[0].endswith(DELIM_LF))
-        self.assertEqual(stats, { DELIM_CR: 0, DELIM_LF: 6, DELIM_CRLF: 0 })
-        
-        stats = { DELIM_CR: 0, DELIM_LF: 0, DELIM_CRLF: 0 }
-        res = read_text_lines("test_mixed.txt", stats)
-        res = list(res)
-        self.assertEqual(len(res), 45)
-        self.assertEqual(type(res[0]), type(b""))
-        self.assertTrue(res[0].endswith(DELIM_CRLF))
-        self.assertEqual(stats, { DELIM_CR: 0, DELIM_LF: 0, DELIM_CRLF: 44 })
+#class TestShell(unittest.TestCase):
+#    """Basic tests.
+#    
+#    Create a temp folder, extract test data there, and CWD to <temp>/test_files:
+#    
+#    We have this structure:
+#    <temp root>/
+#        test_files/
+#            sub1/
+#                (some text and other files)
+#            sub1/
+#                (some text and other files)
+#            (some text and other files)
+#    """
+#    def setUp(self):
+#        # 
+#        if USE_FIXED_FOLDER:
+#            self.temp_path = os.path.join(os.path.expanduser("~"), "tabfix_test_")
+#            if not os.path.exists(self.temp_path):
+#                os.mkdir(self.temp_path)
+#        else:
+#            self.temp_path = tempfile.mkdtemp()
+#        data_path = os.path.dirname(__file__) 
+#        zf = ZipFile(os.path.join(data_path, "test_files.zip"))
+#        zf.extractall(self.temp_path)
+#        self.prev_cwd = os.getcwd()
+#        os.chdir(os.path.join(self.temp_path, "test_files"))
+#
+#        script_folder = os.path.abspath(os.path.join(os.path.dirname(__file__), 
+#                                                        "..", "tabfix"))
+#        if not script_folder in sys.path:
+#            sys.path.append(script_folder) 
+#        self.script_path = os.path.abspath(os.path.join(script_folder, "main.py")) 
+#        
+#    def tearDown(self):
+#        os.chdir(self.prev_cwd)
+#        shutil.rmtree(self.temp_path)
+#        
+#    def test_cmd_help(self):
+##        res = subprocess.call(["tabfix", "-h"])
+#        p = subprocess.Popen(["python", self.script_path, "-h"],
+#                             stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
+#        out, _err = p.communicate()
+#        print(out)
+#        self.assertEqual(p.returncode, 0)
         
 
+#def test_suite():
+#    """Called by 'setup.py test'."""
+#    suite = unittest.TestSuite()
+    
 if __name__ == "__main__":
     print(sys.version)
     unittest.main()
-    
-#    suite = TestSuite()
-##    suite.addTests(TestBasic)
+
+#    suite = unittest.TestSuite()
+#    suite.addTest(TestShell("test_cmd_help"))
 ##    suite.addTest(TestBasic("test_read_text_lines"))
-#    suite.addTest(TestBasic("test_tabbify_txt_flat"))
+##    suite.addTest(TestBasic("test_tabbify_txt_flat"))
 #    unittest.TextTestRunner().run(suite)